fix: f64 config for layer_height/first_layer_height (benchy z=18.8 hull-break regression)
Problem: at z=18.8 (layer 93), the benchy's circular bow hull was
completely missing from the outer wall (280 G1 moves vs 686 in the
no-fix baseline). The prior session's zfix (f64 formula in
layer-planner) eliminated chords at z=9.6/17.2/20.6 but introduced
this regression at z=18.8.

Root cause: double f32-taint in the layer-Z computation chain.
ResolvedConfig stored layer_height: f32 — 0.2_f32 bit pattern is
0.20000000298..., NOT the exact f64 0.2. When to_config_map() widened
back to f64 via f64::from(self.layer_height), the tainted value
0.20000000298... propagated through the WIT boundary. The
layer-planner guest then read it, re-narrowed to f32 (*f as f32),
stored it in f32 fields, and re-widened (as f64) for the formula:
93 x 0.20000000298... = 18.80000028..., which cast to f32
18.80000114... instead of the STL's f32(18.8) = 18.79999924...
The 1.9 um gap exceeded classify_vertex's 1e-6 mm epsilon, so the
on-plane vertex was misclassified as Below -> topology walk dead-ended
-> gap-closer couldn't process -> entire hull loop dropped.

Fix (OrcaSlicer parity): store layer_height and first_layer_height
as f64 in ResolvedConfig, preserving the exact f64 0.2 through to
the layer-planner. The layer-planner stores them as f64 internally
and computes Z in pure f64 (0.2 + 93*0.2 = 18.8 exactly), casting
to f32 only at the WIT LayerProposal.z: f32 boundary — matching
OrcaSlicer's Slicing.cpp:807-867 (coordf_t print_z += height in f64,
float(print_z) at slice_facet's slice_z parameter,
TriangleMeshSlicer.cpp:158).

Changes:
- crates/slicer-ir/src/resolved_config.rs: layer_height/first_layer_height to
  f64, new extract_f64 extractor, to_config_map Eq Hash updates
- modules/core-modules/layer-planner-default/src/lib.rs: all internal
  fields f64, drop redundant as f64 casts
- crates/slicer-scheduler/src/config_resolution.rs: cast cfg.layer_height
  to f32 at support-thinness comparison (display value, not Z input)
- crates/slicer-core/src/algos/paint_segmentation/mod.rs: cast
  cfg.layer_height to f32 at shell-window consumption site
- Test files: type-coercion updates for ResolvedConfig constructors
- crates/slicer-core/src/triangle_mesh_slicer.rs: closure-detection fix
  (benchmark comment, clippy cleanup) + regression test
  f64_layer_z_formula_matches_stl_vertex

Verification:
- cargo check --workspace --all-targets: PASS
- cargo test -p slicer-core: 64 unit + 12 integration — ALL PASS
- cargo test -p slicer-ir: ALL PASS
- cargo test -p layer-planner-default: ALL PASS
- cargo clippy --all-targets -- -D warnings: CLEAN
- Benchy gcode z=18.8: 686 G1 moves (restored), up from 280 (regression)
- Previously-fixed layers (9.6/17.2/20.6/35.4): remain fixed

+++ crates/slicer-core/src/algos/paint_segmentation/mod.rs
@@ -1147,7 +1147,10 @@ pub fn execute_paint_segmentation(
                 cfg.top_shell_layers as usize,
                 cfg.bottom_shell_layers as usize,
                 cfg.line_width,
-                cfg.layer_height,
+                // `cfg.layer_height` is `f64` (Z-formula precision); cast to
+                // `f32` here — shell-window math uses it as a thickness, not
+                // as a Z-plane coordinate, so f32 precision is sufficient.
+                cfg.layer_height as f32,
             ),
             // TODO: when per-object/per-region paint configs are wired through
             // execute_paint_segmentation, prefer the region-specific config
